expose machine state save/load to C API
Saving can be done via lua, but loading can not, because luaengine returns save buffer as a binary string, which can't be loaded back into lua if we can only send lua code. It requires having a custom loader with lua_Reader in sol, which is probably an overkill for now.